backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
httpx>=0.28.0
python-dotenv>=1.0.1
fastembed>=0.7.4
pypdf>=5.9.0
